refactor(mcp): address review coupling
